feat: add shape, image, and text generation services with position and size handling
