+++ backend/src/main/resources/application.properties
@@ -12,3 +12,7 @@ file.upload-dir=./uploads
 
 spring.servlet.multipart.max-file-size=50MB
 spring.servlet.multipart.max-request-size=50MB
+
+logging.level.root=WARN
+logging.level.com.example.webppt=INFO
+logging.level.com.example.webppt.repository=DEBUG
